added colour glow effects to the tiles, and finally started on web3dart. need to create a contract in ganache before further progress can be made.

+++ eightbitbeasts/lib/utils/classes.dart
@@ -12,7 +12,7 @@ class Monster {
 
   String name;
   int id;
-  int grade;
+  double grade;
   String stats;
   int dna;
   Image img;
